Added session 5
For using mapbox-styled basemaps, please use the config-template.js,
and rename it to config.js and register on mapbox.com. Under your
account settings you can create a new mapbox-access-token. Insert it in
the config.js file to see the map. If it doesn’t work or you prefer to
not register on Mapbox, you can use the openstreetmap-code, which I
uncommented in the map.s-file ;)